final project isa ^^

+++ main.sh
@@ -23,11 +23,17 @@ case "${result,,}" in
                     # Extract database name from the user input
                     db_name=$(echo $result | awk '{print $3}')
                     # Create directory with the database name
-                    mkdir -p "databases/$db_name"
-                    # Display dialog with result
-                    echo "Database '$db_name' is created!" | 
-                    zenity --text-info --title="Successful" \
-                    --height 400 --width 600 --font="Arial 20"
+                    if [ -z "$db_name" ]; then
+                    	echo "Database Name Can Not Be Empty!" | 
+                    	zenity --text-info --title="ERROR" \
+                    	--height 400 --width 600 --font="Arial 20"
+                    else 
+                    	mkdir -p "databases/$db_name"
+                    	# Display dialog with result
+                    	echo "Database '$db_name' is created!" | 
+                    	zenity --text-info --title="Successful" \
+                    	--height 400 --width 600 --font="Arial 20"
+                    fi
                     ;;
 "show databases"*)
                     # list all directories in databases dir
@@ -40,16 +46,30 @@ case "${result,,}" in
                     # Extract database name from the user input
                     db_name=$(echo $result | awk '{print $3}')
                     # Drop directory with the database name
-                    rm -df "databases/$db_name"
-                    # Display dialog with result
-                    echo "Database '$db_name' is dropped!" | 
-                    zenity --text-info --title="Successful" \
-                    --height 400 --width 600 --font="Arial 20"
+                    if [ -z "$db_name" ]; then
+                    	echo "Database Name Can Not Be Empty!" | 
+                    	zenity --text-info --title="ERROR" \
+                    	--height 400 --width 600 --font="Arial 20"
+                    elif [ -d databases/$db_name ];then
+                    	rm -df "databases/$db_name"
+                    	# Display dialog with result
+                    	echo "Database '$db_name' is dropped!" | 
+                    	zenity --text-info --title="Successful" \
+                    	--height 400 --width 600 --font="Arial 20"
+                    else 
+                    	echo "Database '$db_name' Not Found!" | 
+                    	zenity --text-info --title="ERROR" \
+                    	--height 400 --width 600 --font="Arial 20"
+                    fi
                     ;;
 "use "*)
                     # Extract database name from the user input
                     db_name=$(echo $result | awk '{print $2}')
-                    if [ -d databases/$db_name ]
+                    if [ -z "$db_name" ]; then
+                    	echo "Database Name Can Not Be Empty!" | 
+                    	zenity --text-info --title="ERROR" \
+                    	--height 400 --width 600 --font="Arial 20"
+                    elif [ -d databases/$db_name ]
                     then
                         echo -e "Connected to $db_name Successfully" |
        		        zenity --text-info --title="Successful" \
@@ -63,14 +83,25 @@ case "${result,,}" in
                     fi
                     ;;
 *)
-		    source ./ddl.sh $result ;
-		    source ./dml.sh $result ;
-		    if [ ! $? -eq 0 ] 
-		    then
+		    if [[ "${result,,}" == "create table "* || "${result,,}" == "drop table "* || "${result,,}" == "show tables"* ]]; then
+		    	source ./ddl.sh $result ;
+		    	if [ ! $? -eq 0 ];then
+		       		# Display dialog with result
+		       		echo "Wrong Statement, $result" |zenity --text-info --title="ERROR" \
+		       		--height 400 --width 600 --font="Arial 20"
+                    	fi
+		    elif [[ "${result,,}" == "insert into "* || "${result,,}" == "select "* || "${result,,}" == "update "* || "${result,,}" == "delete from "* ]]; then
+		    	source ./dml.sh $result ;
+		    	if [ ! $? -eq 0 ];then
+		       		# Display dialog with result
+		       		echo "Wrong Statement, $result" |zenity --text-info --title="ERROR" \
+		       		--height 400 --width 600 --font="Arial 20"
+                    	fi
+		    else
 		       # Display dialog with result
 		       echo "Wrong Statement, $result" |zenity --text-info --title="ERROR" \
 		       --height 400 --width 600 --font="Arial 20"
                     fi
-                     ;;
+                    ;;
     esac
 done
